% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instalar.R
\name{instalar}
\alias{instalar}
\title{Fun\enc{çã}{ca}o Instalar}
\usage{
instalar()
}
\value{
Fun\enc{çã}{ca}o que instala as depend\enc{ê}{e}ncias dos pacotes para o GExpDes
}
\description{
Esta fun\enc{çã}{ca}o faz a instala\enc{çã}{ca}o das depend\enc{ê}{e}ncias dos pacotes para o GExpDes funcionar.
}
\author{
Crystian Paix\enc{ã}{a}o
}
