% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexpdesView.R
\name{gexpdesView}
\alias{gexpdesView}
\title{Func\enc{çã}{ca}o gexpdesView}
\usage{
gexpdesView()
}
\value{
Fun\enc{çã}{ca}o que abre a interface do GExpDes
}
\description{
Esta fun\enc{çã}{ca}o exibe a insterface do GExpDes.
}
\author{
Crystian Paix\enc{ã}{a}o
}
